requirements.txt
streamlit
torch
torchvision
torchaudio
numpy
Pillow
opencv-python

flask
torch
torchvision
pillow
gunicorn
